+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Gamer` Background.
 - Added `Aqua` Deity.
 - Added `Ristarte` Deity.
+- Added `Administrator D` Deity.
 ### Changed
 - Changed `Truck-kun` Deity restriction from `Priest of Balance` to `Angelfire Apostle`.